reorganize functions into _input_sensitivity and rewrite jacobian plotting
