+++ workflows/diagnostics/fv3net/diagnostics/offline/compute.py
@@ -18,11 +18,10 @@ import vcm
 import fv3fit
 from .compute_diagnostics import compute_diagnostics
 from .derived_diagnostics import derived_registry
-from ._plot_input_sensitivity import plot_jacobian, plot_rf_feature_importance
+from ._input_sensitivity import plot_input_sensitivity
 from ._helpers import (
     load_grid_info,
     is_3d,
-    get_variable_indices,
     insert_r2,
     insert_rmse,
     insert_column_integrated_vars,
@@ -39,14 +38,14 @@ logging.basicConfig(handlers=[handler], level=logging.INFO)
 logger = logging.getLogger("offline_diags")
 
 
-# variables that are needed in addition to the model features
+INPUT_SENSITIVITY = "input_sensitivity.png"
 DIAGS_NC_NAME = "offline_diagnostics.nc"
 TRANSECT_NC_NAME = "transect_lon0.nc"
 METRICS_JSON_NAME = "scalar_metrics.json"
 METADATA_JSON_NAME = "metadata.json"
+
 DATASET_DIM_NAME = "dataset"
 DERIVATION_DIM_NAME = "derivation"
-
 DELP = "pressure_thickness_of_atmospheric_layer"
 PREDICT_COORD = "predict"
 TARGET_COORD = "target"
@@ -305,29 +304,12 @@ def main(args):
     ds_diagnostics = ds_diagnostics.update(grid)
 
     # save model senstivity figures- these exclude derived variables
-    base_model = model.base_model if isinstance(model, fv3fit.DerivedModel) else model
-    try:
-        plot_jacobian(
-            base_model,
-            os.path.join(args.output_path, "model_sensitivity_figures"),  # type: ignore
-        )
-    except AttributeError:
-        try:
-            input_feature_indices = get_variable_indices(
-                data=batches[0], variables=base_model.input_variables
-            )
-            plot_rf_feature_importance(
-                input_feature_indices,
-                base_model,
-                os.path.join(args.output_path, "model_sensitivity_figures"),
-            )
-        except AttributeError:
-            logger.info(
-                f"Base model is {type(base_model).__name__}, "
-                "which currently has no feature importance or Jacobian "
-                "calculation implemented."
-            )
-            pass
+    fig_input_sensitivity = plot_input_sensitivity(model, batches[0], args.output_path)
+    with fsspec.open(
+        os.path.join(args.output_dir, "model_sensitivity_figures", INPUT_SENSITIVITY),
+        "wb",
+    ) as f:
+        fig_input_sensitivity.savefig(f)
 
     mapper = _get_data_mapper_if_exists(config)
     if mapper is not None:
